requirements.txt
fastapi
uvicorn[standard]
playwright
openai
python-dotenv
python-multipart